extraindo constantes para arquivo separado
